misc: various cleanup, docs, etc
