+++ .changeset/good-plums-shop.md
@@ -1,0 +1,5 @@
+---
+"gatsby-plugin-fastify": minor
+---
+
+feat: much improved logging
