+++ .changeset/kind-starfishes-change.md
@@ -1,0 +1,5 @@
+---
+"gatsby-plugin-fastify": patch
+---
+
+Fix: SSR/DSG routes throwing non-fatal error when returning file.
