+++ .changeset/odd-days-march.md
@@ -1,0 +1,5 @@
+---
+"gatsby-plugin-fastify": minor
+---
+
+feat: add support for Gatsby 500 error if SSR/DSG throws an error.
