+++ .changeset/stale-rocks-pay.md
@@ -1,0 +1,5 @@
+---
+"gatsby-plugin-fastify": patch
+---
+
+fix: handle situations where dev doesn't provide 404 page by falling back to generic 404.
