+++ packages/gatsby-plugin-fastify/src/plugins/clientPaths.ts
@@ -21,19 +21,13 @@ export const handleClientOnlyPaths: FastifyPluginAsync<{
 
       // This code only works because I've editted the fastify-static implementation to not encodeURI on file names. https://github.com/fastify/fastify-static/issues/234
       // Work around for https://github.com/fastify/fastify/issues/3331
-      // Update, SSR/DSG was implemented without wildcard so this was not an issue. n the future we may need to change this.
-      const fastifyMatchPath = p.matchPath.replace(/\/\*$/, "*");
+      // Update, SSR/DSG was implemented without wildcard so this was not an issue. In the future we may need to change this back if we revert to not sildcarding static routes.
+      // const fastifyMatchPath = p.matchPath.replace(/\/\*$/, "*");
 
-      fastify.get(
-        fastifyMatchPath,
-        {
-          prefixTrailingSlash: "slash",
-        },
-        (_req, reply) => {
-          reply.header("x-gatsby-fastify", `served-by: client-only-routes`);
-          reply.sendFile("index.html", resolve(PATH_TO_PUBLIC, p.path.replace("/", "")));
-        },
-      );
+      fastify.get(p.matchPath, (_req, reply) => {
+        reply.header("x-gatsby-fastify", `served-by: client-only-routes`);
+        reply.sendFile("index.html", resolve(PATH_TO_PUBLIC, p.path.replace("/", "")));
+      });
     }
   }
 };
