+++ packages/gatsby-plugin-fastify/src/plugins/static.ts
@@ -9,7 +9,7 @@ export const handleStatic: FastifyPluginAsync<Partial<FastifyStaticOptions>> = f
   async (fastify, opts) => {
     fastify.register(fastifyStatic, {
       root: resolve(PATH_TO_PUBLIC),
-      wildcard: true,
+      redirect: true,
       setHeaders: (reply, path, _stat) => {
         if (
           isMatch(path, ["**/public/*.@(js|css)", "**/public/static/**"]) &&
